requirements.txt
# Python dependencies for Base AI Project

# Environment variables management
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Disable SSL warnings
urllib3==2.0.7